chore: add an internal cursor to buffer manager
To decode nested frames, we need to not alter the state of the
buffer in case we did not have enough data to decode a full frame.
To achieve that, we need to provide a way for the caller to decide
when it is the good time to consume the decoded data out of the
buffer. So we introduce an internal cursor to remember where we are
but do not alter the buffer itself on decoding.